requirements.txt
pandas==2.1.4
qrcode==7.4.2
Pillow==10.1.0
openpyxl==3.1.2
flask==3.0.0 